requirements.txt
flask 
openai
requests
python-dotenv
groq
python-jose
flask-cors
supabase